+++ src/api/ComputeEngine/ComputeEngineAPI.java
@@ -14,7 +14,7 @@ public interface ComputeEngineAPI {
 	
 	}
 	
-	public static long getSmallestMult(int k) {
+	public static long getSmallestMult(long k) {
 		long n = Long.valueOf(k);
 		long smallMult = 1;
 		long i = 1;
